Add io-ts-promise to README.md community section
I wrote [io-ts-promise](https://github.com/aeirola/io-ts-promise) in order to make it simpler to use io-ts types at promise-based runtime boundary APIs. It was partially inspired by the discussion in #21, where the support for promises was proposed but rejected on the requirement that io-ts needs to stay synchronous.

In case the library would be useful to others as well, I'd propose to add it to the list of community packages.